reface: add json output
add the type of data output by json

Log: add data output type
Issue:
