+++ apps/ll-cli/CMakeLists.txt
@@ -9,5 +9,4 @@ pfl_add_executable(
   ll-cli
   LINK_LIBRARIES
   PUBLIC
-  linglong
-  docopt)
+  linglong)
